% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{abbr_est_names}
\alias{abbr_est_names}
\title{Abbreviate default experiment method names for use in plots}
\usage{
abbr_est_names(x)
}
\arguments{
\item{x}{a vector of method names}
}
\value{
A vector of abbreviated names
}
\description{
Abbreviate default experiment method names for use in plots
}
