% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{centralImputTrNAs}
\alias{centralImputTrNAs}
\title{Handling NAs in train and test}
\usage{
centralImputTrNAs(train, nORp)
}
\arguments{
\item{train}{training data set}

\item{nORp}{minimum percentage of NA values in a row/column for
that row/column to be discarded from training set}
}
\value{
list with an entry for the training and test sets
(in this order), both now with no NA values
}
\description{
Discard columns/rows with too many NAs and 
then impute with central value.
}
