% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{compressRes}
\alias{compressRes}
\title{Compress results from one call to estimates}
\usage{
compressRes(res, ytrain_keep = TRUE, rmAllRaw = F)
}
\arguments{
\item{res}{A list containing full results of one call to estimates}

\item{ytrain_keep}{Should training target values still be kept}

\item{rmAllRaw}{a boolean indicating whether the whole rawRes should
be removed (defaults to FALSE). If TRUE, only \code{train} data will be
substituted by a vector of the training values of the target variable}
}
\value{
A list containing compressed results of one experiment. 
Either all rawRes is removed, or \code{train} substituted by a vector 
containing only the training target values
}
\description{
Compress results from one call to estimates
}
