% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{nb2ggplot}
\alias{nb2ggplot}
\title{From spdep neighbour object to data frame usable with ggplot2}
\usage{
nb2ggplot(nb, stations)
}
\arguments{
\item{nb}{nb object obtained with \code{spdep::dnearneigh} or similar}

\item{stations}{sf object containing geometry points of each location}
}
\value{
data.frame with lon, lat for each point and lon_to and lat_to
as xend and yend for geom_segment
}
\description{
From spdep neighbour object to data frame usable with ggplot2
}
