% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{removeSingleValCol}
\alias{removeSingleValCol}
\title{Remove columns with single unique values from train}
\usage{
removeSingleValCol(train, time, site_id, tgt)
}
\arguments{
\item{train}{training data set}

\item{time}{the name of the column in \code{train} and
\code{test} containing time-stamps}

\item{site_id}{the name of the column in \code{train} and
\code{test} containing location IDs}

\item{tgt}{name of target variable}
}
\description{
Remove columns with single unique values from train
}
