% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_utils.R
\name{summarize_one_exp}
\alias{summarize_one_exp}
\title{Summarize the results of one in-set/out-set experiment}
\usage{
summarize_one_exp(one_exp_res)
}
\arguments{
\item{one_exp_res}{a list containing two slots: \code{out_estRes} containing a 
named vector of metrics estimated in out-set data, and \code{in_estRes} containing a
list of data frames where each column corresponds to a metric and each row to a 
repetition/iteration of an estimator used on in-set data}
}
\value{
A data frame with a first column containing a summary (e.g., the mean) of
metrics measured in the out-set data and further columns containing summaries of 
metrics estimated in the in-set data
}
\description{
Summarize the results of one in-set/out-set experiment
}
\seealso{
\code{\link{run_one_experiment}}
}
